+++ src-tauri/src/lib.rs
@@ -149,7 +149,6 @@ pub fn run() {
             commands::get_storage_file_json,
             commands::write_storage_file_json,
             commands::write_file_bytes,
-            commands::find_suggested_saves,
             commands::set_app_theme,
             commands::validate_recent_saves,
             commands::download_plugin,
@@ -159,6 +158,7 @@ pub fn run() {
             commands::handle_windows_accelerator,
             commands::open_directory,
             commands::open_file_location,
+            saves::find_suggested_saves,
             startup_config::get_data_dir_path,
             startup_config::change_data_dir,
             pkm_storage::load_banks,
